move fee -> cash for now to consolidate the fee logic
