+++ x/cash/init.go
@@ -2,6 +2,7 @@ package cash
 
 import (
 	"github.com/iov-one/weave"
+	"github.com/iov-one/weave/x"
 )
 
 const optKey = "cash"
@@ -41,5 +42,40 @@ func (Initializer) FromGenesis(opts weave.Options, kv weave.KVStore) error {
 			return err
 		}
 	}
+	return nil
+}
+
+// FeeInitializer fulfils the FeeInitializer interface to load data from the genesis
+// file
+type FeeInitializer struct{}
+
+var _ weave.Initializer = (*FeeInitializer)(nil)
+
+// FromGenesis will parse initial account info from genesis and save it to the
+// database
+func (*FeeInitializer) FromGenesis(opts weave.Options, db weave.KVStore) error {
+	var fees []struct {
+		Id  string `json:"id"`
+		Fee x.Coin `json:"fee"`
+	}
+
+	if err := opts.ReadOptions("fees", &fees); err != nil {
+		return err
+	}
+
+	// always default to IOV token
+	for k, v := range fees {
+		v.Fee.Ticker = "IOV"
+		fees[k] = v
+	}
+
+	bucket := NewTransactionFeeBucket()
+	for _, f := range fees {
+		obj := NewTransactionFee(f.Id, f.Fee)
+		if err := bucket.Save(db, obj); err != nil {
+			return err
+		}
+	}
+
 	return nil
 }
