+++ x/cash/init_test.go
@@ -10,6 +10,8 @@ import (
 	"github.com/iov-one/weave/x"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	. "github.com/smartystreets/goconvey/convey"
+
 )
 
 func TestInitState(t *testing.T) {
@@ -79,4 +81,41 @@ func mustCombineCoins(cs ...x.Coin) x.Coins {
 		panic(err)
 	}
 	return s
+}
+
+
+func TestFeeInitializer(t *testing.T) {
+	Convey("Test initializer", t, func() {
+		genesis := `
+		{
+			"fees": [
+				{"id": "currency/tokeninfo",  "fee": {"whole":50,"fractional":1234567}},
+				{"id": "currency/tokeninfo:IOV",  "fee": {"whole":60,"fractional":1234577}}
+			]
+		}`
+		var o weave.Options
+
+		err := json.Unmarshal([]byte(genesis), &o)
+		So(err, ShouldBeNil)
+
+		db := store.MemStore()
+
+		var init FeeInitializer
+		err = init.FromGenesis(o, db)
+		So(err, ShouldBeNil)
+
+		bucket := NewTransactionFeeBucket()
+
+		obj, err := bucket.Get(db, "currency/tokeninfo")
+		So(err, ShouldBeNil)
+		So(obj, ShouldNotBeNil)
+
+		Convey("Match data in the object", func() {
+			fee := obj.Value().(*TransactionFee)
+
+			So(fee.Fee.Fractional, ShouldEqual, 1234567)
+			So(fee.Fee.Whole, ShouldEqual, 50)
+			So(fee.Fee.Ticker, ShouldEqual, "IOV")
+		})
+	})
 }
