--- x/fee/codec.proto
@@ -1,12 +1,0 @@
-syntax = "proto3";
-
-package fee;
-
-import "github.com/iov-one/weave/x/codec.proto";
-
-// TransactionFee represents a fee that has to be deducted for each transaction
-// of a given type. No fee is deducted if this is not configured for a certain transaction
-message TransactionFee {
-  // Fee definition, we ignore ticker and always set it to IOV a discussed
-  x.Coin fee = 1;
-}
